moved JointLimits
